+++ src/middleware/permissionHandler.js
@@ -498,9 +498,10 @@ module.exports = function(router) {
 
     // Check for whitelisted paths.
     if (req.method === 'GET') {
-      const whitelist = ['/health', '/current', '/logout', '/access', '/token'];
+      const whitelist = ['/health', '/current', '/logout', '/access', '/token', '/recaptcha'];
+      const url = req.url.split('?')[0];
       let skip = _.some(whitelist, function(path) {
-        if ((req.url === path) || (req.url === hook.alter('path', path, req))) {
+        if ((url === path) || (url === hook.alter('path', path, req))) {
           return true;
         }