+++ Changelog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/)
 
+## [Unreleased]
+### Added
+ - reCAPTCHA API Endpoint
+
 ## 1.38.0
 ### Added
  - Minimal support for running entire stack in docker-compose